Fix bugs to finaly get working board ...
1. Upper rail for the sense opamp was set to +4V which holds its
output under 1V. This is not enough for control of the MSP which
is expected to be around 1.25V. Changed the upper rail to BOOST_OUT.

2. Signal controlling Q1 battery switch was connected to BOOST_OUT.
This works, but when charger is disconnected boost out capaccitor
discharges relatively slowly and opens Q1 with considerable delay.
Changed the Q1 control to VCHARGE instead.